Now generates a image SVG!.
(first draft) of the following

```
impl FromIterator<String> for Svg {
    fn from_iter<I>(iter: I) -> Self
    where
        I: IntoIterator<Item = String>,
    {
```

This allows the movement of logic from main.rs into svg.rs

At the moment this takes a Vec<String> which is not good for perf.

The logic in Svg now builds up "L X Y" or "l X Y" sub strings.

I put in a "naive" projection which needs to be worked on.

When I say the current struct is not good for performance.. actually

A 1.9MG file "assets/3DBenchy.gcode" can be visualised in 188ms.

```./go.sh > a.svg' benchy

real	0m0.188s
user	0m0.163s
sys	0m0.018s
```

Also

* Removed 'a from Command<'a>

  lifetime was leading to hard to read code.
  Now capturing the dropped code as a u16.

* Added G0 as a non printable MOVE