Bugfix Resolved stringgy models.
The change is perspective :-

"Negative values of E seen in the wild are valid."

The extruder is going into "wipe mode" and is pulling the filament back into
the extruder.

Also Specification Adherenace: G92 – Set Current Position

X and Y are valid parameters and so the virtual head must follow the direction.

Also

Renamed enum CoordPos PositionMode