Performance: Store data payload as a <Cow<'a, [u8]>
Previously the data was always converted from a [u8]
into a String. A conversion is ONLY needed when
displaying a markdown block for example.

Also when a block does not use compression this also
makes use of Cow<'a, T> ability to do a cheap clone.

For example gcode2svg operating on the gears file
the time taken goes from 1m20s down to 53s.

So this is good, but more optimizations are required.