BugFix: G0/G1/G2/G3 subtle parsing bug.
G00 checks must come before G0 checks

Otherwise what comes next a "0" will be interpreted as
as and invalid payload section

This was seem in
-        (alt((tag("G0"), tag("G00"))), space0),
+        (alt((tag("G00"), tag("G0"))), space0),

unit test now work as expected, the test themselves were buggy.

Also added a new snapshot test  - as this was first seen
getting the arc_demo to work.

+    #[test]
+    fn arc_demo() {
+        // SNAPSHOT tests
+        let buffer = include_str!("../../../../assets/arc_demo.gcode");
+        let svg = buffer.lines().map(|l| l.to_string()).collect::<Svg>();
+        assert_debug_snapshot!(svg);
+    }