BugFix: G2/G3 Arc commands are now handled correctly
This issues was advancing theta from start to end
while crossing the zero point.

For examples theta start is in the 4th quadrant
and theta end is in the first quadrant.
The arc has to cross theta = 0/360 point to get to the end.

Removed helper compute_step_params() it was not generic enough